ホーム画面 やるボタンをテーブルの行に統合
